# api.yaml
swagger: "2.0"
info:
  title: Todo API
  version: "1.0.0"
schemes:
  - http
paths:
  /status:
    get:
      summary: staus check
      description: Just a sample string response.
      responses:
        200:
          description: Successful operation
          schema:
            type: string
  /fetch:
    get:
      summary: Get all tasks
      description: Retrieve a list of all tasks from the database.
      security:
        - apiKey: []
      responses:
        200:
          description: Successful operation
          schema:
            type: array
            items:
              $ref: "#/definitions/Task"
        500:
          description: Internal Server Error
  /addTask:
    post:
      summary: Add a task
      description: Add a new task to the todo list.
      security:
        - apiKey: []
      parameters:
        - in: body
          name: task
          description: Task object
          required: true
          schema:
            $ref: "#/definitions/Task"
      responses:
        201:
          description: Task added successfully
        400:
          description: Bad Request - Invalid input
        500:
          description: Internal Server Error
  /completeTask/{id}:
    put:
      summary: Mark a task as complete
      description: Mark a task from the todo list as complete by its ID.
      security:
        - apiKey: []
      parameters:
        - in: path
          name: id
          description: ID of the task to mark as complete
          required: true
          type: integer
      responses:
        200:
          description: Task marked as complete successfully
        404:
          description: Task not found
        400:
          description: Bad Request - Invalid input
        500:
          description: Internal Server Error
  /deleteTask/{id}:
    delete:
      summary: Delete a task by ID
      description: Delete a task from the todo list by its ID.
      security:
        - apiKey: []
      parameters:
        - in: path
          name: id
          description: ID of the task to delete
          required: true
          type: integer
      responses:
        200:
          description: Task deleted successfully
        404:
          description: Task not found
        400:
          description: Bad Request - Invalid input
        500:
          description: Internal Server Error

definitions:
  Task:
    type: object
    properties:
      id:
        type: integer
        description: Task ID
      tasks:
        type: string
        description: Task description
      complete:
        type: integer
        description: Task completion status (0 for incomplete, 1 for complete)

securityDefinitions:
  apiKey:
    type: apiKey
    in: header
    name: X-API-KEY

security:
  - apiKey: []